Factored out PhotoZipper class
